added tests required for hw 3 and updated repo_miner to handle some edge cases and updated README with basic usage info
