+++ pyproject.toml
@@ -7,5 +7,6 @@ requires-python = ">=3.12"
 dependencies = [
     "PyGithub",
     "pandas",
-    "pytest"
+    "pytest",
+    "pytest-recording"
 ]
