+++ tests/dummies.py
@@ -1,0 +1,57 @@
+# --- Helpers for dummy GitHub API objects ---
+
+class DummyAuthor:
+    def __init__(self, name, email, date):
+        self.name = name
+        self.email = email
+        self.date = date
+
+class DummyCommitCommit:
+    def __init__(self, author, message):
+        self.author = author
+        self.message = message
+
+class DummyCommit:
+    def __init__(self, sha, author, email, date, message):
+        self.sha = sha
+        self.commit = DummyCommitCommit(DummyAuthor(author, email, date), message)
+
+class DummyUser:
+    def __init__(self, login):
+        self.login = login
+
+class DummyIssue:
+    def __init__(self, id_, number, title, user, state, created_at, closed_at, comments, is_pr=False):
+        self.id = id_
+        self.number = number
+        self.title = title
+        self.user = DummyUser(user)
+        self.state = state
+        self.created_at = created_at
+        self.closed_at = closed_at
+        self.comments = comments
+        # attribute only on pull requests
+        self.pull_request = DummyUser("pr") if is_pr else None
+
+class DummyRepo:
+    def __init__(self, commits, issues):
+        self._commits = commits
+        self._issues = issues
+
+    def get_commits(self):
+        return self._commits
+
+    def get_issues(self, state="all"):
+        # filter by state
+        if state == "all":
+            return self._issues
+        return [i for i in self._issues if i.state == state]
+
+class DummyGithub:
+    def __init__(self, token, auth=None):
+        assert token == "fake-token"
+    def get_repo(self, repo_name):
+        # ignore repo_name; return repo set in test fixture
+        return self._repo
+    def close(self):
+        pass # no-op for dummy
