+++ tests/test_repo.py
@@ -5,93 +5,58 @@ import pandas as pd
 import pytest
 from datetime import datetime, timedelta
 from src.repo_miner import fetch_commits, fetch_issues, merge_and_summarize
-
-# --- Helpers for dummy GitHub API objects ---
-
-class DummyAuthor:
-    def __init__(self, name, email, date):
-        self.name = name
-        self.email = email
-        self.date = date
-
-class DummyCommitCommit:
-    def __init__(self, author, message):
-        self.author = author
-        self.message = message
-
-class DummyCommit:
-    def __init__(self, sha, author, email, date, message):
-        self.sha = sha
-        self.commit = DummyCommitCommit(DummyAuthor(author, email, date), message)
-
-class DummyUser:
-    def __init__(self, login):
-        self.login = login
-
-class DummyIssue:
-    def __init__(self, id_, number, title, user, state, created_at, closed_at, comments, is_pr=False):
-        self.id = id_
-        self.number = number
-        self.title = title
-        self.user = DummyUser(user)
-        self.state = state
-        self.created_at = created_at
-        self.closed_at = closed_at
-        self.comments = comments
-        # attribute only on pull requests
-        self.pull_request = DummyUser("pr") if is_pr else None
-
-class DummyRepo:
-    def __init__(self, commits, issues):
-        self._commits = commits
-        self._issues = issues
-
-    def get_commits(self):
-        return self._commits
-
-    def get_issues(self, state="all"):
-        # filter by state
-        if state == "all":
-            return self._issues
-        return [i for i in self._issues if i.state == state]
-
-class DummyGithub:
-    def __init__(self, token):
-        assert token == "fake-token"
-    def get_repo(self, repo_name):
-        # ignore repo_name; return repo set in test fixture
-        return self._repo
-
-@pytest.fixture(autouse=True)
-def patch_env_and_github(monkeypatch):
-    # Set fake token
-    monkeypatch.setenv("GITHUB_TOKEN", "fake-token")
-    # Patch Github class
-    # TODO
-
-# Helper global placeholder
-gh_instance = DummyGithub("fake-token")
+from tests.dummies import DummyAuthor, DummyCommitCommit, DummyCommit, DummyUser, DummyIssue, DummyRepo, DummyGithub
 
 # --- Tests for fetch_commits ---
-# An example test case
+class TestsWithDummies:
+    
+    # Helper instance
+    gh_instance = DummyGithub("fake-token")
+    
+    @pytest.fixture(autouse=True)
+    def patch_env_and_github(self, monkeypatch):
+        # Set fake token
+        monkeypatch.setenv("GITHUB_TOKEN", "fake-token")
+        # Patch Github class'
+        monkeypatch.setattr("src.repo_miner.Github", lambda auth=None: self.gh_instance)
+        
+    
+    # An example test case
+    def test_fetch_commits_basic(self, monkeypatch):
+        # Setup dummy commits
+        now = datetime.now()
+        commits = [
+            DummyCommit("sha1", "Alice", "a@example.com", now, "Initial commit\nDetails"),
+            DummyCommit("sha2", "Bob", "b@example.com", now - timedelta(days=1), "Bug fix")
+        ]
+        self.gh_instance._repo = DummyRepo(commits, [])
+        df = fetch_commits("any/repo")
+        assert list(df.columns) == ["sha", "author", "email", "date", "message"]
+        assert len(df) == 2
+        assert df.iloc[0]["message"] == "Initial commit"
+
+# --- Tests that hit real GitHub API (will be simulated with vcrpy) ---
+@pytest.mark.vcr
 def test_fetch_commits_basic(monkeypatch):
-    # Setup dummy commits
-    now = datetime.now()
-    commits = [
-        DummyCommit("sha1", "Alice", "a@example.com", now, "Initial commit\nDetails"),
-        DummyCommit("sha2", "Bob", "b@example.com", now - timedelta(days=1), "Bug fix")
-    ]
-    gh_instance._repo = DummyRepo(commits, [])
-    df = fetch_commits("any/repo")
+    """ Basic fetch on well known small repo """
+    df = fetch_commits("octocat/Hello-World")
+    assert list(df.columns) == ["sha", "author", "email", "date", "message"]
+    assert len(df) == 3
+    assert df.iloc[0]["message"] == "Merge pull request #6 from Spaceghost/patch-1"
+
+@pytest.mark.vcr
+def test_fetch_commits_limit(monkeypatch):
+    """ More commits than max_commits. Test that fetch_commits respects the max_commits limit. """
+    df = fetch_commits("octocat/Hello-World", max_commits=2)
     assert list(df.columns) == ["sha", "author", "email", "date", "message"]
     assert len(df) == 2
-    assert df.iloc[0]["message"] == "Initial commit"
-
-def test_fetch_commits_limit(monkeypatch):
-    # More commits than max_commits
-    # TODO： Test that fetch_commits respects the max_commits limit.
-    pass
-
+    assert df.iloc[0]["message"] == "Merge pull request #6 from Spaceghost/patch-1"
+        
+@pytest.mark.vcr
 def test_fetch_commits_empty(monkeypatch):
-    # TODO: Test that fetch_commits returns empty DataFrame when no commits exist.
-    pass
+    """ Test that fetch_commits returns empty DataFrame when no max commits exist. """
+    # Uses a cassette modified so that it returns 0 commits
+    df = fetch_commits("octocat/Hello-World")
+    assert list(df.columns) == ["sha", "author", "email", "date", "message"]
+    assert len(df) == 0
+    assert df.empty
